Update D2-json2yolo.py
新增功能：为负样本创建空白的txt标签

+++ Python/常用脚本/codes/D2-json2yolo.py
@@ -1,7 +1,6 @@
 """
     json转yolo的txt
 """
-
 import os
 import cv2
 import json
@@ -10,6 +9,7 @@ import tqdm
 import logging
 import os
 import datetime
+from tabulate import tabulate
 
 
 """============================ 需要修改的地方 ==================================="""
@@ -23,11 +23,18 @@ IMAGE_TYPE = '.jpg'  # 图片类型
 class_dict = {"cat": 0, 
               "dog": 1}
 
+CONVERT_CLASS_DICT = False  # 是否需要把类别字典的key和value翻转一下?
+
+# 是否为负样本创建txt文件
+CREATE_TXT_4_NEG = True
+
 DECIMAL_PLACES = 6  # 标签保留几位小数, 默认为6
 OVERRIDE = True  # 是否要覆盖已存在txt文件
 CLIP_OUT_OF_BOUNDARY = True  # 对于 xmin 或 ymin < 0 的情况，直接截断为 0; 对于 xmax 或 ymax > 图片尺寸的情况，直接截断图片最大尺寸
 VERBOSE = False  # 终端不会打印日志了（日志仍会生成）
 """==============================================================================="""
+
+class_dict = {v: k for k, v in class_dict.items()} if CONVERT_CLASS_DICT else ... # 翻转字典
 
 "---------------------------------------日志---------------------------------------"
 # 获取当前时间
@@ -56,23 +63,30 @@ WARNING_NUM = 0
 ERROR_NUM = 0
 ERROR_LOGGER = dict()
 ERROR_POINT_NUM = 0
+NEG_NUM = 0
 "---------------------------"
 del not_image_list
 
-_str = (f" 图片路径: \033[1;33m{IMAGE_PATH}\033[0m"
-        f"\n json路径: \033[1;33m{ANNOTATION_PATH}\033[0m"
-        f"\n txt保存路径: \033[1;33m{LABELS_PATH}\033[0m"
-        f"\n {IMAGE_TYPE}图片数量: \033[1;33m{len(image_list)}\033[0m"
-        f"\n 💡 不是 {IMAGE_TYPE} 图片数量: \033[1;33m{NOT_WANNA_IMAGES_NUM}\033[0m"
-        f"\n 需要转换的json文件数量: \033[1;33m{TOTAL_NUM}\033[0m"
-        f"\n 日志保存路径: \033[1;33m{log_file_path}\033[0m"
-        f"\n\t💡 是否要覆盖: \033[1;33m{OVERRIDE}\033[0m"
-        f"\n\t💡 是否对越界坐标进行截断: \033[1;33m{CLIP_OUT_OF_BOUNDARY}\033[0m"
-        f"\n\t💡 日志是否在终端显示(不影响日志保存): \033[1;33m{VERBOSE}\033[0m"
-        f"\n\n请输入 \033[1;31m'yes'\033[0m 继续，输入其他停止"
-        )
+_str = [
+    ["图片路径", IMAGE_PATH],
+    ["json路径", ANNOTATION_PATH],
+    ["txt保存路径", LABELS_PATH],
+    ["", ""],
+    [f"{IMAGE_TYPE}图片数量", len(image_list)],
+    ["💡 不是图片数量", NOT_WANNA_IMAGES_NUM],
+    ["需要转换的json文件数量", TOTAL_NUM],
+    ["", ""],
+    ["💡 是否为负样本创建txt文件", CREATE_TXT_4_NEG],
+    ["💡 是否要覆盖", OVERRIDE],
+    ["💡 是否对越界坐标进行截断", CLIP_OUT_OF_BOUNDARY],
+    ["", ""],
+    ["日志保存路径", log_file_path],
+    ["💡 日志是否在终端显示", VERBOSE],
+]
 
-print(_str)
+_str = tabulate(_str, headers=["PARAMs", "VALUE"], tablefmt="pipe")
+print(f"{_str}\n\n"
+      f"请输入 'yes' 继续，输入其他停止")
     
 _INPUT = input()
 if _INPUT != "yes":
@@ -111,8 +125,10 @@ def recalc_coordinate(xmin, ymin, xmax, ymax, json_path, WARNING_NUM=0, logger=N
         xmin, ymin = xcenter - width / 2, ycenter - height / 2
         xmax, ymax = xcenter + width / 2, ycenter + height / 2
 
-        logger.warning(f"坐标错误, 已重新计算!\n\t"
-                       f"[[{_xmin}, {_ymin}], [{_xmax}, {_ymax}]] -> \n\t[[{xmin}, {ymin}], [{xmax}, {ymax}]]:\n\t"
+        logger.warning(f"坐标错误, 已重新计算!\n\t\t"
+                       f"[({_xmin}, {_ymin}), ({_xmax}, {_ymax})]\n"
+                       f"---new--->\n\t\t"
+                       f"[({xmin}, {ymin}), ({xmax}, {ymax})]:\n\t"
                        f"{json_path}") if logger else ...
 
     return xmin, ymin, xmax, ymax, WARNING_NUM
@@ -139,25 +155,25 @@ def check_coordinate_validity(xmin, ymin, xmax, ymax,
         _xmin = xmin
         if CLIP_OUT_OF_BOUNDARY:
             xmin = 0.0
-        logger.error(msg=f"[xmin({_xmin}) < 0]\n\tNow xmin={xmin}\n\t{json_path}")
+        logger.error(msg=f"[xmin({_xmin}) < 0] ---new---> xmin={xmin}\n\t{json_path}")
         ERROR_NUM += 1
     if ymin < 0: 
         _ymin = ymin
         if CLIP_OUT_OF_BOUNDARY:
             ymin = 0.0
-        logger.error(msg=f"[ymin({_ymin}) < 0]\n\tNow ymin={ymin}\n\t{json_path}")
+        logger.error(msg=f"[ymin({_ymin}) < 0] ---new---> ymin={ymin}\n\t{json_path}")
         ERROR_NUM += 1
     if xmax < 0: 
         _xmax = xmax
         if CLIP_OUT_OF_BOUNDARY:
             xmax = 0.0
-        logger.error(msg=f"[xmax({_xmax}) < 0]\n\tNow xmax={xmax}\n\t{json_path}")
+        logger.error(msg=f"[xmax({_xmax}) < 0] ---new---> xmax={xmax}\n\t{json_path}")
         ERROR_NUM += 1
     if ymax < 0: 
         _ymax = ymax
         if CLIP_OUT_OF_BOUNDARY:
             ymax = 0.0
-        logger.error(msg=f"[ymax({_ymax}) < 0]\n\tNow ymax={ymax}\n\t{json_path}")
+        logger.error(msg=f"[ymax({_ymax}) < 0] ---new---> ymax={ymax}\n\t{json_path}")
         ERROR_NUM += 1
 
     # 如果出现越界
@@ -165,25 +181,25 @@ def check_coordinate_validity(xmin, ymin, xmax, ymax,
         _xmin = xmin
         if CLIP_OUT_OF_BOUNDARY:
             xmin = float(img_width)
-        logger.error(msg=f"[xmin > img_width({_xmin} > {img_width})]\n\tNow xmin={xmin}\n\t{json_path}")
+        logger.error(msg=f"[xmin > img_width({_xmin} > {img_width})] ---new---> xmin={xmin}\n\t{json_path}")
         ERROR_NUM += 1
     if ymin > img_height: 
         _ymin = ymin
         if CLIP_OUT_OF_BOUNDARY:
             ymin = float(img_height)
-        logger.error(msg=f"[ymin > img_height({_ymin} > {img_height})]\n\tNow ymin={ymin}\n\t{json_path}")
+        logger.error(msg=f"[ymin > img_height({_ymin} > {img_height})] ---new---> ymin={ymin}\n\t{json_path}")
         ERROR_NUM += 1
     if xmax > img_width: 
         _xmax = xmax
         if CLIP_OUT_OF_BOUNDARY:
             xmax = float(img_width)
-        logger.error(msg=f"[xmax > img_width({_xmax} > {img_width})]\n\tNow xmax={xmax}\n\t{json_path}")
+        logger.error(msg=f"[xmax > img_width({_xmax} > {img_width})] ---new---> xmax={xmax}\n\t{json_path}")
         ERROR_NUM += 1
     if ymax > img_height: 
         _ymax = ymax
         if CLIP_OUT_OF_BOUNDARY:
             ymax = float(img_height)
-        logger.error(msg=f"[ymax > img_height({_ymax} > {img_height})]\n\tNow ymax={ymax}\n\t{json_path}")
+        logger.error(msg=f"[ymax > img_height({_ymax} > {img_height})] ---new---> ymax={ymax}\n\t{json_path}")
         ERROR_NUM += 1
     
     return xmin, ymin, xmax, ymax, ERROR_NUM
@@ -209,17 +225,22 @@ for _, json_name in enumerate(annotation_list):
         result = json.load(fr)
 
     # 读取图片的宽高信息
+    if not os.path.exists(img_path):  # 如果对应的图片不存在
+        ERROR_NUM += 1
+        logger.error(msg=f"[图片不存在!] {img_path}\n\t{json_path}")
+        progress_bar.update()
+        continue
     img = cv2.imread(img_path)
     img_height, img_width = img.shape[0:2]  # 👌
     
-    # 获取所有 object 的信息 -> e.g. [{'label': 'dough_hambuger', 'text': '', 'points': [[619.1836734693877, 347.14285714285717], [657.9591836734694, 377.7551020408164]], 'group_id': None, 'shape_type': 'rectangle', 'flags': {}}]
+    # 获取所有 object 的信息 -> e.g. [{'label': 'dough_hambuger', 'text': '', 'points': [(619.1836734693877, 347.14285714285717), (657.9591836734694, 377.7551020408164)], 'group_id': None, 'shape_type': 'rectangle', 'flags': {}}]
     all_object_info = result['shapes']
     
     # 创建 txt 文件并写入内容
     with open(txt_path, 'w') as target_file:
         for idx, obj_info in enumerate(all_object_info):
-            label = class_dict[obj_info['label']]  # 读取当前obj的类别
-            points = np.array(obj_info['points'])  # 读取当前obj的位置 -> e.g. [[619.18367347 347.14285714] [657.95918367 377.75510204]]
+            label = label_dict[obj_info['label']]  # 读取当前obj的类别
+            points = np.array(obj_info['points'])  # 读取当前obj的位置 -> e.g. [(619.18367347 347.14285714] [657.95918367 377.75510204]]
 
             # 防止出现一个点
             _pt_len = len(points)
@@ -272,13 +293,30 @@ for _, json_name in enumerate(annotation_list):
     progress_bar.update(1)
 progress_bar.close()
 
+# 最后考虑到可能有部分图片没有标签，所以我们主动给他们创建负样本
+if CREATE_TXT_4_NEG:
+    for image_name in tqdm.tqdm(image_list):
+        pre, ext = os.path.splitext(image_name)
+        txt_path = os.path.join(LABELS_PATH, pre) + '.txt'  # 最终保存的txt文件完整路径
+
+        # 如果目标txt文件存在则跳过
+        if os.path.exists(txt_path):
+            continue
+        
+        # 如果目标 txt 文件不存在，创建空的 txt 文件
+        with open(txt_path, 'w', encoding='utf-8') as txt_file:
+            image_path = os.path.join(IMAGE_PATH, image_name)
+            NEG_NUM += 1
+            logger.warning(msg=f"[负样本] [{NEG_NUM}] {image_path}")
+
 _str = (f"json2yolo已完成，详情如下：\n\t"
         f"👌成功: {SUCCEED_NUM}/{TOTAL_NUM}\n\t"
         f"👌跳过: {SKIP_NUM}/{TOTAL_NUM}\n\t"
+        f"负样本数量: {NEG_NUM}\n\t"
         f"Object数量: {OBJ_NUM}\n\t"
         f"每张图片平均Object数量: {OBJ_NUM/TOTAL_NUM:.2f}\n\t"
         f"坏点的数量为: {ERROR_POINT_NUM}\n\n"
-        f"结果保存在: {LABELS_PATH}"
+        f"结果保存在: {LABELS_PATH}\n"
         f"日志保存在: {log_file_path}")
 
 logger.info(_str)
